refactored code, moved add and update files, added more checks to update, and and delete files, started styling
