+++ wedding-planner/src/components/dashboard/DashBoardHeader.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import {Link} from 'react-router-dom';
 
 
 
@@ -6,10 +7,9 @@ const DashBoardHeader = (props) =>{
     return(
         <header>
             <nav>
-                <div>Name</div>
                 <div>Wedding Planner</div>
-                <div># of events ({props.events.length})</div>
-                <div>Logout</div>
+                <div># of events ({props.count})</div>
+                <div><Link to={`/dashboard/${props.id}/add`}>Add Event</Link></div>
             </nav>
         </header>
     )
